requirements.txt
streamlit 
langchain-community
google-genai 
requests
langchain-google-genai